Create Destroyable Module

Adding wrappers to a shiny module to enable an ease of dynamically adding and removing modules within a shiny application.

makeModuleUIDestroyable(module_fn, wrapper = shiny::div)

makeModuleServerDestroyable(module_fn)

Arguments

module_fn

The server-side part of the module

wrapper

If the module is a `shiny::tagList()`, then an HTML tag will be wrapped by an HTML tag so that a shiny.destroy attribute can be attached

Examples

library(shiny)

# UI
basicModuleUI <- function(id) {
  ns <- NS(id)
  actionButton("click", "Increase")
}

destroyableModuleUI <- makeModuleUIDestroyable(basicModuleUI)

# Server-side
basicMoudleServer <- function(id) {
  moduleServer(id, function(input, output, session) {
    rv <- reactiveVal()
    observeEvent(input$click, rv(input$click))
  })
}

destroyableModuleServer <- makeModuleServerDestroyable(basicMoudleServer)

# Shiny Application
ui <- fluidPage(
  destroyableModuleUI(id = "test"),
  actionButton("destroy", "Destroy module"),
  textOutput("reactive_value")
)

server <- function(input, output, session) {
  top_rv <- reactiveVal()

  reactive_value <- destroyableModuleServer("test")
  observeEvent(reactive_value(), top_rv(reactive_value()))

  output$reactive_value <- renderText(top_rv())

  observeEvent(input$destroy, destroyModule("test"))
}